Add logo to header
Logo is a .png because that's what was provided. It's implemented as a
pseudoelement of the h1 because I didn't want a decorative image
clogging up the DOM. The styling was finnicky to get right, but I think
it's the more semantically reasonable choice, because the h1 text does
go with the image.